src/requirements.txt
matplotlib
numpy
Pillow
flask
flask_httpauth
json_tricks
